requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1

# LangChain core + integrations
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-pinecone>=0.2.0

# Vector DB client
pinecone>=6.0.0

# PDF loading
pypdf>=4.2.0

# Optional local embedding fallback (not required if using OpenAI embeddings)
# sentence-transformers>=2.2.2
